Extract controller and Epic Games helpers
