+++ games/alanwake2/alanwake2.py
@@ -15,7 +15,7 @@ sys.path.insert(1, PARENT_DIRECTORY)
 from harness_utils.artifacts import ArtifactManager, ArtifactType
 from harness_utils.input import press_n_times
 from harness_utils.ocr_service import find_word
-from harness_utils.misc import find_eg_game_version
+from harness_utils.epic_games import find_eg_game_version
 from harness_utils.output import setup_logging, write_report_json
 from harness_utils.process import terminate_process
